+++ CHANGELOG.md
@@ -78,6 +78,7 @@
 * Drop `EveOnline::XML::AccountCharacters`. ESI doesn't support this.
 * Drop `EveOnline::XML::ApiKeyInfo`
 * Drop `EveOnline::XML::AccountStatus`. ESO doesn't support this.
+* Add `EveOnline::ESI::Models::Standing`
 
 **v0.12.0**